gcp-configure/docker/pyproject.toml
[tool.poetry]
name = "marathon-cloud-run-sample-docker"
version = "0.1.0"
description = ""
authors = ["threecourse"]

[tool.poetry.dependencies]
python = "3.7.*"
numpy = "1.16.4"
pandas = "0.24.2"
google-cloud-logging = "1.15.0"
google-cloud-storage = "1.28.1"
google-api-python-client = "2.2.0"
requests = "2.22.0"
gunicorn = "20.0.4"
Flask = "1.1.1"

[tool.poetry.dev-dependencies]
ipython = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
